+++ src/Libraries/SmartStore.Core/Domain/Tax/TaxSettings.cs
@@ -137,11 +137,5 @@ namespace SmartStore.Core.Domain.Tax
         /// Gets or sets a value indicating whether we should notify a store owner when a new VAT number is submitted
         /// </summary>
         public bool EuVatEmailAdminWhenNewVatSubmitted { get; set; }
-
-		/// <summary>
-		/// Gets or sets a value indicating the ESD tax class identifier required
-		/// to calculate VAT for electronic services conform to EU VAT regulations (2015)
-		/// </summary>
-		public int? ElectronicServiceTaxClassId { get; set; }
     }
 }