+++ src/Libraries/SmartStore.Services/Catalog/CopyProductService.cs
@@ -170,6 +170,7 @@ namespace SmartStore.Services.Catalog
 				IsShipEnabled = product.IsShipEnabled,
 				IsFreeShipping = product.IsFreeShipping,
 				AdditionalShippingCharge = product.AdditionalShippingCharge,
+				IsEsd = product.IsEsd,
 				IsTaxExempt = product.IsTaxExempt,
 				TaxCategoryId = product.TaxCategoryId,
 				ManageInventoryMethod = product.ManageInventoryMethod,